Docker-based test environment GitHub Actions CI
