+++ .github/workflows/integration-tests.yml
@@ -1,0 +1,47 @@
+---
+
+name: CI
+
+on:
+  push:
+    branches:
+      - master
+  pull_request:
+    branches:
+      - master
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    name: CI
+
+    strategy:
+      fail-fast: false
+      matrix:
+        node-version: [15.x, 14.x, 12.x, 10.x]
+
+    steps:
+      - name: Checkout Repository
+        uses: actions/checkout@v2
+        with:
+          fetch-depth: 1
+
+      - name: Setup Node ${{ matrix.node-version }}
+        uses: actions/setup-node@v2
+        with:
+          always-auth: false
+          node-version: ${{ matrix.node-version }}
+
+      - name: Start Solr
+        run: npm run solr:start
+
+      - name: Run npm install
+        run: npm install
+
+      - name: Run Tests
+        run: npm run test
+        env:
+          CI: true
+
+      - name: Stop Solr
+        run: npm run solr:stop
